Onyx Capital/FinansDashboard/backend/requirements.txt
flask
flask-cors
flask-jwt-extended
psycopg2-binary
pandas
openpyxl